+++ ballerina/init.bal
@@ -21,5 +21,5 @@ isolated function init() {
 }
 
 isolated function setModule() = @java:Method {
-    'class: "org.ballerinalang.oracledb.utils.ModuleUtils"
+    'class: "io.ballerina.stdlib.oracledb.utils.ModuleUtils"
 } external;
