+++ gradle.properties
@@ -10,7 +10,7 @@ githubSpotbugsVersion=4.0.5
 githubJohnrengelmanShadowVersion=5.2.0
 underCouchDownloadVersion=4.0.4
 researchgateReleaseVersion=2.8.0
-ballerinaGradlePluginVersion=0.9.1
+ballerinaGradlePluginVersion=0.10.1
 
 ballerinaLangVersion=2.0.0-beta.2.1
