+++ native/src/main/java/io/ballerina/stdlib/oracledb/Constants.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.oracledb;
+package io.ballerina.stdlib.oracledb;
 
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BString;
