+++ native/src/main/java/io/ballerina/stdlib/oracledb/utils/ProcedureCallResultUtils.java
@@ -16,27 +16,28 @@
  *  under the License.
  */
 
-package org.ballerinalang.oracledb.utils;
+package io.ballerina.stdlib.oracledb.utils;
 
 import io.ballerina.runtime.api.values.BObject;
-import org.ballerinalang.oracledb.parameterprocessor.OracleDBResultParameterProcessor;
+import io.ballerina.stdlib.oracledb.parameterprocessor.OracleDBResultParameterProcessor;
 import org.ballerinalang.sql.parameterprocessor.DefaultResultParameterProcessor;
 
 /**
- * This class provides functionality to call `sql:RecordIteratorUtils` with a custom `ResultParameterProcessor` object.
+ * This class provides functionality to call `sql:ProcedureCallResult` with a custom `ResultParameterProcessor` object.
  *
  * @since 0.1.0
  */
-public class RecordIteratorUtils {
+public class ProcedureCallResultUtils {
 
     /**
-     * Call `sql:RecordIteratorUtils` with a custom `ResultParameterProcessor` object.
+     * Calls sql:ProcedureCallResult` with a custom `ResultParameterProcessor` object.
      * @param customResultIterator module specific resultIterator BObject
-     * @param iterator the record that needs to be iterated
-     * @return next result of the iterator
+     * @param callResult call result that needs to be iterated
+     * @return next query result
      */
-    public static Object nextResult(BObject customResultIterator, BObject iterator) {
+    public static Object getNextQueryResult(BObject customResultIterator, BObject callResult) {
         DefaultResultParameterProcessor resultParameterProcessor = OracleDBResultParameterProcessor.getInstance();
-        return org.ballerinalang.sql.utils.RecordIteratorUtils.nextResult(iterator, resultParameterProcessor);
+        return org.ballerinalang.sql.utils.ProcedureCallResultUtils.getNextQueryResult(
+                callResult, resultParameterProcessor);
     }
 }
